requirements.txt
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
copier==9.7.1
cryptography==44.0.2
cycler==0.12.1
databricks-cli==0.18.0
databricks-connect==15.1.3
databricks-sdk==0.61.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
dunamai==1.24.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.11
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
flake8==7.2.0
fonttools==4.58.4
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipyvue==1.11.2
ipyvuetify==1.11.3
ipywidgets==8.1.7
-e git+https://github.com/tpomavil46/itv_asset_tree.git@061dfcb2ee95c0c0b0906365737603b9a500373a#egg=itv_asset_tree
jedi==0.19.2
Jinja2==3.1.6
jinja2-ansible-filters==1.3.2
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lxml==5.4.0
makefun==1.15.6
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.79.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.2.0
pluggy==1.5.0
plumbum==1.9.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.1
py4j==0.10.9.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyspark==3.5.5
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.0
questionary==2.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
rsa==4.9.1
seaborn==0.13.2
seeq==66.12.3.20250318
seeq-spy==194.0
setuptools-scm==8.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
streamlit==1.47.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
wrapt==1.17.2
